Explicitly handle vertical/horizontal mouse wheel scrolling in Delphi's code editor
This is now needed for scrolling with the TrackPoint of the Lenovo Thinkpad, at least on newer models, such as the Thinkpad T14 Gen 4. The driver on older models probably already converted the mouse wheel messages to WM_VSCROLL/WM_HSCROLL itself (but it wasn't possible anymore to verify this). Now, we need to do this conversion in GExperts.